+++ lib/l10n/modules/settings/settings_en.arb
@@ -34,8 +34,8 @@
   "preferences_autoFillLoginDesc": "Remember credentials and auto-fill on login",
   "preferences_clipboardTopicLinkDetection": "Detect clipboard topic links",
   "preferences_clipboardTopicLinkDetectionDesc": "Check for Linux.do topic links in the clipboard when returning to the app and ask before opening",
-  "clipboardTopicLink_detected": "Topic link found in clipboard",
-  "clipboardTopicLink_open": "Open",
+  "preferences_clipboardTopicLink_detected": "Topic link found in clipboard",
+  "preferences_clipboardTopicLink_open": "Open",
   "preferences_autoPanguSpacing": "Auto CJK spacing",
   "preferences_autoPanguSpacingDesc": "Auto-insert spaces between CJK and Latin text",
   "preferences_basic": "Basic",
