+++ lib/l10n/modules/settings/settings_zh.arb
@@ -51,8 +51,8 @@
   "preferences_autoFillLoginDesc": "记住账号密码，登录时自动填充",
   "preferences_clipboardTopicLinkDetection": "自动识别剪贴板话题链接",
   "preferences_clipboardTopicLinkDetectionDesc": "回到应用时检测剪贴板中的 Linux.do 话题链接，并在底部询问是否打开",
-  "clipboardTopicLink_detected": "检测到剪贴板中的话题链接",
-  "clipboardTopicLink_open": "打开",
+  "preferences_clipboardTopicLink_detected": "检测到剪贴板中的话题链接",
+  "preferences_clipboardTopicLink_open": "打开",
   "preferences_autoPanguSpacing": "自动混排优化",
   "preferences_autoPanguSpacingDesc": "输入时自动插入中英文混排空格",
   "preferences_basic": "基础",
