+++ lib/l10n/modules/settings/settings_zh_HK.arb
@@ -34,8 +34,8 @@
   "preferences_autoFillLoginDesc": "記住賬號密碼，登錄時自動填充",
   "preferences_clipboardTopicLinkDetection": "自動識別剪貼板話題鏈接",
   "preferences_clipboardTopicLinkDetectionDesc": "回到應用時檢測剪貼板中的 Linux.do 話題鏈接，並在底部詢問是否打開",
-  "clipboardTopicLink_detected": "檢測到剪貼板中的話題鏈接",
-  "clipboardTopicLink_open": "打開",
+  "preferences_clipboardTopicLink_detected": "檢測到剪貼板中的話題鏈接",
+  "preferences_clipboardTopicLink_open": "打開",
   "preferences_autoPanguSpacing": "自動混排優化",
   "preferences_autoPanguSpacingDesc": "輸入時自動插入中英文混排空格",
   "preferences_basic": "基礎",
