+++ lib/l10n/modules/settings/settings_zh_TW.arb
@@ -34,8 +34,8 @@
   "preferences_autoFillLoginDesc": "記住賬號密碼，登入時自動填充",
   "preferences_clipboardTopicLinkDetection": "自動偵測剪貼簿話題連結",
   "preferences_clipboardTopicLinkDetectionDesc": "回到應用時偵測剪貼簿中的 Linux.do 話題連結，並在底部詢問是否開啟",
-  "clipboardTopicLink_detected": "偵測到剪貼簿中的話題連結",
-  "clipboardTopicLink_open": "開啟",
+  "preferences_clipboardTopicLink_detected": "偵測到剪貼簿中的話題連結",
+  "preferences_clipboardTopicLink_open": "開啟",
   "preferences_autoPanguSpacing": "自動混排最佳化",
   "preferences_autoPanguSpacingDesc": "輸入時自動插入中英文混排空格",
   "preferences_basic": "基礎",
