+++ lib/main.dart
@@ -20,6 +20,7 @@ import 'services/discourse/discourse_service.dart';
 import 'providers/app_state_refresher.dart';
 import 'services/highlighter_service.dart';
 import 'widgets/common/notification_icon_button.dart';
+import 'widgets/common/clipboard_topic_link_snack_content.dart';
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'services/network/cookie/android_cdp_feature.dart';
 import 'services/network/cookie/csrf_token_service.dart';
@@ -496,114 +497,6 @@ class MainApp extends ConsumerWidget {
   }
 }
 
-class _ClipboardTopicLinkSnackContent extends StatelessWidget {
-  final String message;
-  final String actionLabel;
-  final VoidCallback onOpen;
-  final VoidCallback onDismiss;
-
-  const _ClipboardTopicLinkSnackContent({
-    required this.message,
-    required this.actionLabel,
-    required this.onOpen,
-    required this.onDismiss,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final colorScheme = theme.colorScheme;
-    final isDark = theme.brightness == Brightness.dark;
-    final borderRadius = BorderRadius.circular(8);
-    final backgroundColor = Color.alphaBlend(
-      colorScheme.primary.withValues(alpha: isDark ? 0.14 : 0.06),
-      colorScheme.surfaceContainerHigh,
-    );
-
-    return Align(
-      alignment: Alignment.bottomCenter,
-      child: ConstrainedBox(
-        constraints: const BoxConstraints(maxWidth: 560),
-        child: Material(
-          color: backgroundColor,
-          elevation: 8,
-          shadowColor: colorScheme.shadow.withValues(
-            alpha: isDark ? 0.38 : 0.18,
-          ),
-          surfaceTintColor: colorScheme.surfaceTint,
-          borderRadius: borderRadius,
-          child: DecoratedBox(
-            decoration: BoxDecoration(
-              borderRadius: borderRadius,
-              border: Border.all(
-                color: colorScheme.outlineVariant.withValues(
-                  alpha: isDark ? 0.32 : 0.6,
-                ),
-              ),
-            ),
-            child: Padding(
-              padding: const EdgeInsets.fromLTRB(12, 10, 8, 10),
-              child: Row(
-                children: [
-                  Container(
-                    width: 32,
-                    height: 32,
-                    decoration: BoxDecoration(
-                      color: colorScheme.primaryContainer.withValues(
-                        alpha: isDark ? 0.62 : 0.78,
-                      ),
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    alignment: Alignment.center,
-                    child: Icon(
-                      Icons.link_rounded,
-                      size: 19,
-                      color: colorScheme.onPrimaryContainer,
-                    ),
-                  ),
-                  const SizedBox(width: 12),
-                  Expanded(
-                    child: Text(
-                      message,
-                      maxLines: 2,
-                      overflow: TextOverflow.ellipsis,
-                      style: theme.textTheme.bodyMedium?.copyWith(
-                        color: colorScheme.onSurface,
-                        fontWeight: FontWeight.w500,
-                      ),
-                    ),
-                  ),
-                  const SizedBox(width: 8),
-                  TextButton(
-                    onPressed: onOpen,
-                    style: TextButton.styleFrom(
-                      minimumSize: const Size(56, 48),
-                      padding: const EdgeInsets.symmetric(horizontal: 12),
-                    ),
-                    child: Text(actionLabel),
-                  ),
-                  IconButton(
-                    onPressed: onDismiss,
-                    tooltip: MaterialLocalizations.of(
-                      context,
-                    ).closeButtonTooltip,
-                    icon: const Icon(Icons.close_rounded),
-                    iconSize: 20,
-                    style: IconButton.styleFrom(
-                      foregroundColor: colorScheme.onSurfaceVariant,
-                      minimumSize: const Size(48, 48),
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
 Future<void> _syncSlangLocale(Locale? locale) async {
   if (locale == null) {
     await LocaleSettings.useDeviceLocale();
@@ -662,7 +555,11 @@ class _MainPageState extends ConsumerState<MainPage>
 
       // 初始化 Deep Link 服务
       DeepLinkService.instance.initialize(context);
-      unawaited(_runStartupUiTasks());
+      unawaited(
+        _runStartupUiTasks().catchError((Object e, StackTrace s) {
+          debugPrint('[MainPage] 启动 UI 任务失败: $e\n$s');
+        }),
+      );
     });
     // 监听登录失效事件
     _authErrorSub = ref.listenManual<AsyncValue<String>>(authErrorProvider, (
@@ -954,9 +851,9 @@ class _MainPageState extends ConsumerState<MainPage>
     messenger.hideCurrentSnackBar();
     final controller = messenger.showSnackBar(
       SnackBar(
-        content: _ClipboardTopicLinkSnackContent(
-          message: context.l10n.clipboardTopicLink_detected,
-          actionLabel: context.l10n.clipboardTopicLink_open,
+        content: ClipboardTopicLinkSnackContent(
+          message: context.l10n.preferences_clipboardTopicLink_detected,
+          actionLabel: context.l10n.preferences_clipboardTopicLink_open,
           onOpen: () {
             markPromptedOnce();
             messenger.hideCurrentSnackBar();
@@ -972,7 +869,6 @@ class _MainPageState extends ConsumerState<MainPage>
         duration: const Duration(seconds: 8),
         elevation: 0,
         padding: EdgeInsets.zero,
-        persist: false,
         margin: EdgeInsets.only(
           left: 16,
           right: 16,
