+++ lib/services/clipboard_topic_link_service.dart
@@ -28,7 +28,19 @@ class ClipboardTopicLinkService {
     caseSensitive: false,
   );
 
+  /// 进程内已提示过的链接 hash。
+  /// 与持久化的 [lastPromptedHashPrefsKey] 配合使用：持久化只记最近一次，
+  /// 而本集合用于在同一进程内挡掉历史提示链接，避免来回切换前后台时反复弹。
+  /// 上限保护避免长生命周期累积。
+  static const int _maxSeenHashes = 64;
   final Set<int> _seenHashes = <int>{};
+
+  void _rememberSeen(int hash) {
+    _seenHashes.add(hash);
+    if (_seenHashes.length > _maxSeenHashes) {
+      _seenHashes.remove(_seenHashes.first);
+    }
+  }
 
   Future<ClipboardTopicLinkCandidate?> checkClipboard({
     required bool enabled,
@@ -58,23 +70,24 @@ class ClipboardTopicLinkService {
     ClipboardTopicLinkCandidate candidate, {
     SharedPreferences? prefs,
   }) async {
-    _seenHashes.add(candidate.normalizedHash);
+    _rememberSeen(candidate.normalizedHash);
     await prefs?.setInt(lastPromptedHashPrefsKey, candidate.normalizedHash);
   }
 
   ClipboardTopicLinkCandidate? findFirstTopicLink(String text) {
     for (final match in _linuxDoUrlRegex.allMatches(text)) {
       final rawUrl = _trimTrailingPunctuation(match.group(0)!);
-      if (!_hasValidLeadingBoundary(text, match.start, rawUrl)) continue;
+      if (!_hasValidLeadingBoundary(text, match.start)) continue;
 
       final uri = _parseUri(rawUrl);
       if (uri == null || !_isAllowedHost(uri.host)) continue;
       if (!_isSupportedTopicPath(uri.path)) continue;
       if (DiscourseUrlParser.parseTopic(uri.path) == null) continue;
 
-      final normalizedUrl = _normalizeUrl(uri);
+      final normalizedUri = _normalize(uri);
+      final normalizedUrl = normalizedUri.toString();
       return ClipboardTopicLinkCandidate(
-        uri: Uri.parse(normalizedUrl),
+        uri: normalizedUri,
         normalizedUrl: normalizedUrl,
         normalizedHash: _stableHash(normalizedUrl),
       );
@@ -83,12 +96,13 @@ class ClipboardTopicLinkService {
     return null;
   }
 
-  static bool _hasValidLeadingBoundary(String text, int start, String rawUrl) {
+  static bool _hasValidLeadingBoundary(String text, int start) {
     if (start == 0) return true;
     if (_looksEmbeddedInAnotherUrl(text, start)) return false;
 
     final previous = text.codeUnitAt(start - 1);
-    if (rawUrl.startsWith('//') && previous == 0x3a) return false;
+    // 前一个字符是冒号一律拒绝，覆盖 ://linux.do 与 mailto:linux.do 等情况
+    if (previous == 0x3a) return false;
     if (previous == 0x3d || previous == 0x26) return false;
 
     return !_isAsciiLetterOrDigit(previous) &&
@@ -185,11 +199,11 @@ class ClipboardTopicLinkService {
     return parsed != null && parsed > 0;
   }
 
-  static String _normalizeUrl(Uri uri) {
+  static Uri _normalize(Uri uri) {
     final scheme = uri.scheme.toLowerCase();
     final host = uri.host.toLowerCase();
     final keepPort = uri.hasPort && !_isDefaultPort(scheme, uri.port);
-    final normalizedUri = Uri(
+    return Uri(
       scheme: scheme,
       host: host,
       port: keepPort ? uri.port : null,
@@ -197,8 +211,6 @@ class ClipboardTopicLinkService {
       query: uri.hasQuery ? uri.query : null,
       fragment: uri.hasFragment ? uri.fragment : null,
     );
-
-    return normalizedUri.toString();
   }
 
   static bool _isDefaultPort(String scheme, int port) {
